Add subscription package
This commit adds subscription package which is responsible for deciding
whether a subscription/invoice can be paid or not.

The endpoint fails every other requests that it receives. As far as I
understood from Antaeus README, payment provider should fail 50% of the
requests, but Antaeus stops making the requests (exists loop) once one
invoice is failed to be paid, so without changing the client code
(Antaeus) we cannot make 50% of *all* the requests fail.